Add a temporary processor for DIV and SPAN
