+++ packages/roosterjs-content-model/lib/domToModel/context/defaultProcessors.ts
@@ -3,6 +3,7 @@ import { ElementProcessor } from '../../publicTypes/context/ElementProcessor';
 import { fontProcessor } from '../processors/fontProcessor';
 import { knownElementProcessor } from '../processors/knownElementProcessor';
 import { tableProcessor } from '../processors/tableProcessor';
+import { tempContainerProcessor } from '../processors/tempContainerProcessor';
 
 /**
  * @internal
@@ -10,10 +11,12 @@ import { tableProcessor } from '../processors/tableProcessor';
 export const defaultProcessorMap: Record<string, ElementProcessor> = {
     B: knownElementProcessor,
     BR: brProcessor,
+    DIV: tempContainerProcessor,
     EM: knownElementProcessor,
     FONT: fontProcessor,
     I: knownElementProcessor,
     S: knownElementProcessor,
+    SPAN: tempContainerProcessor,
     STRIKE: knownElementProcessor,
     STRONG: knownElementProcessor,
     SUB: knownElementProcessor,
