--- packages/roosterjs-content-model/lib/domToModel/processors/generalBlockProcessor.ts
@@ -1,14 +1,0 @@
-import { addBlock } from '../../modelApi/common/addBlock';
-import { containerProcessor } from './containerProcessor';
-import { createGeneralBlock } from '../../modelApi/creators/createGeneralBlock';
-import { ElementProcessor } from '../../publicTypes/context/ElementProcessor';
-
-/**
- * @internal
- */
-export const generalBlockProcessor: ElementProcessor = (group, element, context) => {
-    const block = createGeneralBlock(element);
-
-    addBlock(group, block);
-    containerProcessor(block, element, context);
-};
