--- packages/roosterjs-content-model/test/domToModel/processors/generalBlockProcessorTest.ts
@@ -1,41 +1,0 @@
-import * as containerProcessor from '../../../lib/domToModel/processors/containerProcessor';
-import * as createGeneralBlock from '../../../lib/modelApi/creators/createGeneralBlock';
-import { ContentModelGeneralBlock } from '../../../lib/publicTypes/block/group/ContentModelGeneralBlock';
-import { createContentModelDocument } from '../../../lib/modelApi/creators/createContentModelDocument';
-import { createDomToModelContext } from '../../../lib/domToModel/context/createDomToModelContext';
-import { DomToModelContext } from '../../../lib/publicTypes/context/DomToModelContext';
-import { generalBlockProcessor } from '../../../lib/domToModel/processors/generalBlockProcessor';
-
-describe('generalBlockProcessor', () => {
-    let context: DomToModelContext;
-
-    beforeEach(() => {
-        spyOn(containerProcessor, 'containerProcessor');
-        context = createDomToModelContext();
-    });
-
-    it('Process a DIV element', () => {
-        const doc = createContentModelDocument(document);
-        const div = document.createElement('div');
-        const block: ContentModelGeneralBlock = {
-            blockType: 'BlockGroup',
-            blockGroupType: 'General',
-            element: div,
-            blocks: [],
-        };
-
-        spyOn(createGeneralBlock, 'createGeneralBlock').and.returnValue(block);
-        generalBlockProcessor(doc, div, context);
-
-        expect(doc).toEqual({
-            blockType: 'BlockGroup',
-            blockGroupType: 'Document',
-            blocks: [block],
-            document: document,
-        });
-        expect(createGeneralBlock.createGeneralBlock).toHaveBeenCalledTimes(1);
-        expect(createGeneralBlock.createGeneralBlock).toHaveBeenCalledWith(div);
-        expect(containerProcessor.containerProcessor).toHaveBeenCalledTimes(1);
-        expect(containerProcessor.containerProcessor).toHaveBeenCalledWith(block, div, context);
-    });
-});
