+++ packages/roosterjs-content-model/test/domToModel/processors/generalProcessorTest.ts
@@ -1,17 +1,44 @@
 import * as containerProcessor from '../../../lib/domToModel/processors/containerProcessor';
+import * as createGeneralBlock from '../../../lib/modelApi/creators/createGeneralBlock';
 import * as createGeneralSegment from '../../../lib/modelApi/creators/createGeneralSegment';
+import { ContentModelGeneralBlock } from '../../../lib/publicTypes/block/group/ContentModelGeneralBlock';
 import { ContentModelGeneralSegment } from '../../../lib/publicTypes/segment/ContentModelGeneralSegment';
 import { createContentModelDocument } from '../../../lib/modelApi/creators/createContentModelDocument';
 import { createDomToModelContext } from '../../../lib/domToModel/context/createDomToModelContext';
 import { DomToModelContext } from '../../../lib/publicTypes/context/DomToModelContext';
-import { generalSegmentProcessor } from '../../../lib/domToModel/processors/generalSegmentProcessor';
+import { generalProcessor } from '../../../lib/domToModel/processors/generalProcessor';
 
-describe('generalSegmentProcessor', () => {
+describe('generalProcessor', () => {
     let context: DomToModelContext;
 
     beforeEach(() => {
-        context = createDomToModelContext();
         spyOn(containerProcessor, 'containerProcessor');
+        context = createDomToModelContext();
+    });
+
+    it('Process a DIV element', () => {
+        const doc = createContentModelDocument(document);
+        const div = document.createElement('div');
+        const block: ContentModelGeneralBlock = {
+            blockType: 'BlockGroup',
+            blockGroupType: 'General',
+            element: div,
+            blocks: [],
+        };
+
+        spyOn(createGeneralBlock, 'createGeneralBlock').and.returnValue(block);
+        generalProcessor(doc, div, context);
+
+        expect(doc).toEqual({
+            blockType: 'BlockGroup',
+            blockGroupType: 'Document',
+            blocks: [block],
+            document: document,
+        });
+        expect(createGeneralBlock.createGeneralBlock).toHaveBeenCalledTimes(1);
+        expect(createGeneralBlock.createGeneralBlock).toHaveBeenCalledWith(div);
+        expect(containerProcessor.containerProcessor).toHaveBeenCalledTimes(1);
+        expect(containerProcessor.containerProcessor).toHaveBeenCalledWith(block, div, context);
     });
 
     it('Process a SPAN element', () => {
@@ -28,7 +55,7 @@ describe('generalSegmentProcessor', () => {
 
         spyOn(createGeneralSegment, 'createGeneralSegment').and.returnValue(segment);
 
-        generalSegmentProcessor(doc, span, context);
+        generalProcessor(doc, span, context);
 
         expect(doc).toEqual({
             blockType: 'BlockGroup',
@@ -53,7 +80,7 @@ describe('generalSegmentProcessor', () => {
         const span = document.createElement('span');
         context.segmentFormat = { a: 'b' } as any;
 
-        generalSegmentProcessor(doc, span, context);
+        generalProcessor(doc, span, context);
 
         expect(doc).toEqual({
             blockType: 'BlockGroup',
